requirements.txt
llama-index==0.14.10
llama-index-vector-stores-faiss==0.5.1

sentence-transformers==2.6.1
faiss-cpu==1.7.4

ollama==0.1.8

imapclient==3.0.1
pyzmail36==1.0.5
python-dotenv==1.0.1
